Passenv

Description

passenv is a tool and Haskell library for fetching secrets from a password store managed by pass and adding them to your environment.

See https://github.com/splintah/passenv for more information.

Documentation

passenv :: PassenvOptions -> Builder IO Env -> IO void

Run the default passenv function:

Get the process' arguments which describe the command to execute. Throws an ArgumentsError when not enough arguments were supplied.

Run the given builder, which builds an Env in the IO monad.

Execute the command in the new environment.

This function will never finish, because the passenv process will be replaced by the executed process.

passenvWith

Arguments

:: String
Command to run.

-> [String]
Arguments for the command.

-> PassenvOptions

-> Builder IO Env

-> IO void

Does the same as passenv, but does not get the arguments of the process, and instead uses the command and arguments that are passed to the function.

data PassenvOptions

Options for passenv.

There is a Default instance for PassenvOptions; you can use def to get the default options.

Constructors

PassenvOptions

Fields

passenvSearchPath :: Bool

Search the current path for the command.

See executeFile for more information.

Default: True.

data PassenvException

An exception for passenv functions.

Constructors

ArgumentsError String
Wrong or not enough arguments were supplied. The String describes the intended usage.

Re-exports

From Passenv.Pass

pass :: (MonadIO m, MonadThrow m, MonadCatch m) => String -> Builder (ReaderT PasswordFile m) a -> Builder m a

pass f b: read a file named f from the password store, parse it as a PasswordFile, run the builder b that builds values of type a in a context with a PasswordFile, and return a Builder for a.

passKV :: (MonadIO m, MonadThrow m, MonadCatch m) => String -> Builder (ReaderT KeyValueFile m) a -> Builder m a

The same as passKVWith, but use the default options.

passKVWith :: (MonadIO m, MonadThrow m, MonadCatch m) => KeyValueFileOptions -> String -> Builder (ReaderT KeyValueFile m) a -> Builder m a

passKV opts f b: read a file named f from the password store, parse it as a KeyValueFile with the options in opts, run the builder b that builds values of type a in a context with a KeyValueFile, and return a Builder for a.

passWithParser :: (MonadIO m, MonadThrow m, MonadCatch m, Show e, ShowErrorComponent e, Typeable e) => Parsec e String r -> String -> Builder (ReaderT r m) a -> Builder m a

passWithParser p f b: read a file named f from the password store, parse it using p which returns a value of type r, run the builder b that builds values of type a in a context with a value of type r, and return a Builder for a.

From Passenv.Env

env :: Applicative m => [(Builder m Var, Builder m Var)] -> Builder m Env

Create an Env Builder from a list of pairs of Var Builders.

From Passenv.Builder

class HasPassword r where

r contains a password value.

Methods

password :: MonadReader r m => Builder m String

A Builder which returns the password value.

class HasKeyValues r where

r contains key-value pairs.

Minimal complete definition

valueMaybe

Methods

valueMaybe :: MonadReader r m => String -> Builder m (Maybe String)

Get the value of the given key, returning Nothing if it is not found.

value :: (MonadReader r m, MonadThrow m) => String -> Builder m String

Get the value of the given key, throwing a KeyValueException if it is not found.

From external modules

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

Methods

def :: a

The default value for this type.