Passenv.Builder

Description

A builder for monoidal values in a monadic context.

This module contains the Builder type that is used for building the Env, but also for the underlying Vars. And it can be used for any Monoidal value, and in any Monadic context (an Applicative context is often enough).

IsString

This module also exports an IsString instance for Builders, which is useful when building Strings or other types with an IsString instance.

For example, if we have a builder

stringBuilder :: Applicative m => Builder m String

then we can compose it with a String literal:

stringBuilder <> "some string" :: Applicative m => Builder m String

For this you must enable the OverloadedStrings language extension, by adding the following to the top of your file:

Builder

data Builder m a

A builder for monoidal values in a monadic context.

Builders are Monoids if a is a Monoid and m is an Applicative, and can be composed with (<>). Builders are also Monads if m is a Monad.

mkBuilder :: m a -> Builder m a

Make a Builder from a monadic action.

runBuilder :: Builder m a -> m a

Run a Builder, returning the underlying monadic action.

Has classes

class HasPassword r where

r contains a password value.

Methods

password :: MonadReader r m => Builder m String

A Builder which returns the password value.

class HasKeyValues r where

r contains key-value pairs.

Minimal complete definition

valueMaybe

Methods

valueMaybe :: MonadReader r m => String -> Builder m (Maybe String)

Get the value of the given key, returning Nothing if it is not found.

value :: (MonadReader r m, MonadThrow m) => String -> Builder m String

Get the value of the given key, throwing a KeyValueException if it is not found.

data KeyValueException

An exception for operations on a value which HasKeyValues.

Constructors

KeyNotPresentError String
The given key is not present.