Passenv.Env

Description

Environment builders.

The env Builder is used to build an environment.

Documentation

type Var = String

An environment variable. This type is used for both the variable's name and the variable's value.

newtype Env

An environment. This is a list of pairs of Vars.

Constructors

Env

Fields

getEnv :: [(Var, Var)]

env :: Applicative m => [(Builder m Var, Builder m Var)] -> Builder m Env

Create an Env Builder from a list of pairs of Var Builders.