Passenv.Pass

Description

Functions for interacting with the password store.

Documentation

pass :: (MonadIO m, MonadThrow m, MonadCatch m) => String -> Builder (ReaderT PasswordFile m) a -> Builder m a

pass f b: read a file named f from the password store, parse it as a PasswordFile, run the builder b that builds values of type a in a context with a PasswordFile, and return a Builder for a.

passKV :: (MonadIO m, MonadThrow m, MonadCatch m) => String -> Builder (ReaderT KeyValueFile m) a -> Builder m a

The same as passKVWith, but use the default options.

passKVWith :: (MonadIO m, MonadThrow m, MonadCatch m) => KeyValueFileOptions -> String -> Builder (ReaderT KeyValueFile m) a -> Builder m a

passKV opts f b: read a file named f from the password store, parse it as a KeyValueFile with the options in opts, run the builder b that builds values of type a in a context with a KeyValueFile, and return a Builder for a.

passWithParser :: (MonadIO m, MonadThrow m, MonadCatch m, Show e, ShowErrorComponent e, Typeable e) => Parsec e String r -> String -> Builder (ReaderT r m) a -> Builder m a

passWithParser p f b: read a file named f from the password store, parse it using p which returns a value of type r, run the builder b that builds values of type a in a context with a value of type r, and return a Builder for a.

passContents :: String -> IO String

passContents name gets the contents of the file named name in the password store.

The contents are retrieved by executing the pass command with name as its argument.

data PassException

An exception that occurs when interacting with the password store.

Constructors

PassProcessError
The pass process returned with an error.

Fields

Int

Process exit code.

String

stderr of the process.

forall e.Exception e => PassErrorWithFile
An (unrelated) exception was thrown when interacting with a file in the password store.

Fields

String

The name of the file.

e

The exception

class HasPassword r where

r contains a password value.

Methods

password :: MonadReader r m => Builder m String

A Builder which returns the password value.

class HasKeyValues r where

r contains key-value pairs.

Minimal complete definition

valueMaybe

Methods

valueMaybe :: MonadReader r m => String -> Builder m (Maybe String)

Get the value of the given key, returning Nothing if it is not found.

value :: (MonadReader r m, MonadThrow m) => String -> Builder m String

Get the value of the given key, throwing a KeyValueException if it is not found.