Passenv.Pass.KeyValue

Description

A file format with a password and key-value pairs.

Documentation

data KeyValueFile

A key-value file is a password file format that contains:

On the first line: a password.

On all other lines: key-value pairs in the format key: value. The spaces before and after the : will be ignored; this means that keys cannot contain spaces, and values cannot start with spaces.

This file format is often used by browser extensions for pass.

This file format can be parsed using the keyValueFile and keyValueFileWith parsers.

The parser also supports comments, starting with whatever character you like. You can configure this character using kvFileOptionsComment. Comments must appear on their own line, so the comment character can appear inside values, as can be seen in the following example.

Example

myverygoodpassword
username: myusername
email:user@example.com
# A line comment, and an empty line:

key         : value
test: this is not # a comment

With the default options, parsing this file will return:

KeyValueFile
  { kvFilePassword = "myverygoodpassword"
  , kvFileMap = 'Map.fromList'
    [ ("email","user@example.com")
    , ("key","value")
    , ("test", "this is not # a comment")
    , ("username","myusername")
    ]
  }

Constructors

KeyValueFile

Fields

kvFilePassword :: String

The password.

kvFileMap :: Map String String

The key-value pairs.

data KeyValueFileOptions

The options for parsing a KeyValueFile.

This type has a Default instance, so you can use def as the default options.

Constructors

KeyValueFileOptions

Fields

kvFileOptionsComment :: Maybe Char

The character that starts comments. If Nothing, comments will not be parsed. If Just c for some c :: Char, the character c will start comments.

See KeyValueFile for more information about comments.

Default: Just '#'.

Parsers

keyValueFile :: Parser KeyValueFile

Parse a KeyValueFile with default KeyValueFileOptions.

keyValueFileWith :: KeyValueFileOptions -> Parser KeyValueFile

Parse a KeyValueFile with the given KeyValueFileOptions.

type Parser = Parsec Void String

The type of the parser.

Re-exports

class HasPassword r where

r contains a password value.

Methods

password :: MonadReader r m => Builder m String

A Builder which returns the password value.

class HasKeyValues r where

r contains key-value pairs.

Methods

valueMaybe :: MonadReader r m => String -> Builder m (Maybe String)

Get the value of the given key, returning Nothing if it is not found.