Passenv.Pass.Password

Description

A file format with a password on the first line.

Documentation

data PasswordFile

A PasswordFile is a file that contains a password on the first line. The rest of the contents will not be parsed.

This file format can be parsed using the passwordFile parser.

Example

myverygoodpassword
key: value
Foo

# Bar

Parsing this file will return:

PasswordFile
  { pwFilePassword = "myverygoodpassword"
  }

Constructors

PasswordFile

Fields

pwFilePassword :: String

The password.

Parser

passwordFile :: Parser PasswordFile

Parse a passwordFile.

type Parser = Parsec Void String

The type of the parser.

Re-exports

class HasPassword r where

r contains a password value.

Methods

password :: MonadReader r m => Builder m String

A Builder which returns the password value.