Passenv.Util

Description

A file format with a password on the first line.

Documentation

line :: (MonadParsec e s m, Stream s, Token s ~ Char) => m (Tokens s)

Parses a line with an optional end of line.

isEol :: Char -> Bool

Returns whether the given Char is an end of line character.

isEol c = c == '\n' || c == '\r'